+++ meerkat/columns/abstract.py
@@ -52,9 +52,17 @@ class AbstractColumn(
         self,
         data: Sequence = None,
         collate_fn: Callable = None,
+        formatter: Callable = None,
         *args,
         **kwargs,
     ):
+        """
+
+        Args:
+            data (Sequence, optional): [description]. Defaults to None.
+            collate_fn (Callable, optional): [description]. Defaults to None.
+            formatter (Callable, optional): . Defaults to None.
+        """
         # Assign to data
         self._set_data(data)
 
@@ -62,6 +70,10 @@ class AbstractColumn(
             collate_fn=collate_fn,
             *args,
             **kwargs,
+        )
+
+        self._formatter = (
+            formatter if formatter is not None else self._get_default_formatter()
         )
 
         # Log creation
@@ -97,7 +109,7 @@ class AbstractColumn(
     @classmethod
     def _state_keys(cls) -> set:
         """List of attributes that describe the state of the object."""
-        return {"_collate_fn"}
+        return {"_collate_fn", "_formatter"}
 
     def _get_cell(self, index: int, materialize: bool = True) -> Any:
         """Get a single cell from the column.
@@ -239,8 +251,14 @@ class AbstractColumn(
     def _repr_cell_(self, index) -> object:
         raise NotImplementedError
 
-    def _get_formatter(self) -> callable:
+    def _get_default_formatter(self) -> Callable:
+        # can't implement this as a class level property because then it will treat
+        # the formatter as a method
         return None
+
+    @property
+    def formatter(self) -> Callable:
+        return self._formatter
 
     def _repr_pandas_(self, max_rows: int = None) -> pd.Series:
         if max_rows is None:
@@ -258,8 +276,7 @@ class AbstractColumn(
         else:
             col = pd.Series([self._repr_cell(idx) for idx in range(len(self))])
 
-        formatter = self._get_formatter()
-        return col, formatter
+        return col, self.formatter
 
     def _repr_html_(self, max_rows: int = None):
         # pd.Series objects do not implement _repr_html_
