+++ meerkat/columns/image_column.py
@@ -1,15 +1,10 @@
 from __future__ import annotations
 
 import logging
-import os
-import urllib.request
-from typing import Collection, Sequence
-from urllib.error import HTTPError
-from urllib.parse import urlparse
+from typing import Callable
 
-from meerkat.columns.abstract import AbstractColumn
-from meerkat.columns.lambda_column import LambdaCell, LambdaColumn
-from meerkat.columns.pandas_column import PandasSeriesColumn
+from meerkat.columns.file_column import FileColumn
+from meerkat.display import image_file_formatter
 from meerkat.tools.lazy_loader import LazyLoader
 
 folder = LazyLoader("torchvision.datasets.folder")
@@ -17,57 +12,7 @@ folder = LazyLoader("torchvision.datasets.folder")
 logger = logging.getLogger(__name__)
 
 
-class ImageLoaderMixin:
-    def fn(self, filepath: str):
-        absolute_path = (
-            os.path.join(self.base_dir, filepath)
-            if self.base_dir is not None
-            else filepath
-        )
-        image = self.loader(absolute_path)
-
-        if self.transform is not None:
-            image = self.transform(image)
-        return image
-
-
-class ImageCell(ImageLoaderMixin, LambdaCell):
-    def __init__(
-        self,
-        transform: callable = None,
-        loader: callable = None,
-        data: str = None,
-        base_dir: str = None,
-    ):
-        self.loader = self.default_loader if loader is None else loader
-        self.transform = transform
-        self._data = data
-        self.base_dir = base_dir
-
-    @property
-    def absolute_path(self):
-        return (
-            os.path.join(self.base_dir, self.data)
-            if self.base_dir is not None
-            else self.data
-        )
-
-    def __eq__(self, other):
-        return (
-            (other.__class__ == self.__class__)
-            and (self.data == other.data)
-            and (self.transform == other.transform)
-            and (self.loader == other.loader)
-        )
-
-    def __repr__(self):
-        transform = getattr(self.transform, "__qualname__", repr(self.transform))
-        dirs = self.data.split("/")
-        short_path = ("" if len(dirs) <= 2 else ".../") + "/".join(dirs[-2:])
-        return f"ImageCell({short_path}, transform={transform})"
-
-
-class ImageColumn(ImageLoaderMixin, LambdaColumn):
+class ImageColumn(FileColumn):
     """A column where each cell represents an image stored on disk. The underlying data
     is a `PandasSeriesColumn` of strings, where each string is the path to an image.
     The column materializes the images into memory when indexed. If the column is
@@ -97,97 +42,10 @@ class ImageColumn(ImageLoaderMixin, LambdaColumn):
 
     """
 
-    def __init__(
-        self,
-        data: Sequence[str] = None,
-        transform: callable = None,
-        loader: callable = None,
-        base_dir: str = None,
-        *args,
-        **kwargs,
-    ):
-
-        if not isinstance(data, PandasSeriesColumn):
-            data = PandasSeriesColumn(data)
-        super(ImageColumn, self).__init__(data, *args, **kwargs)
-        self.loader = self.default_loader if loader is None else loader
-        self.transform = transform
-        self.base_dir = base_dir
-
-    def _create_cell(self, data: object) -> ImageCell:
-        return ImageCell(
-            data=data,
-            loader=self.loader,
-            transform=self.transform,
-            base_dir=self.base_dir,
-        )
-
-    @classmethod
-    def from_filepaths(
-        cls,
-        filepaths: Sequence[str],
-        loader: callable = None,
-        transform: callable = None,
-        base_dir: str = None,
-        *args,
-        **kwargs,
-    ):
-        return cls(
-            data=filepaths,
-            loader=loader,
-            transform=transform,
-            base_dir=base_dir,
-            *args,
-            **kwargs,
-        )
+    @staticmethod
+    def _get_default_formatter() -> Callable:
+        return image_file_formatter
 
     @classmethod
     def default_loader(cls, *args, **kwargs):
         return folder.default_loader(*args, **kwargs)
-
-    @classmethod
-    def _state_keys(cls) -> Collection:
-        return (super()._state_keys() | {"transform", "loader", "base_dir"}) - {"fn"}
-
-    def _set_state(self, state: dict):
-        state["base_dir"] = state.get("base_dir", None)  # backwards compatibility
-        super()._set_state(state)
-
-    def is_equal(self, other: AbstractColumn) -> bool:
-        return (
-            (other.__class__ == self.__class__)
-            and (self.loader == other.loader)
-            and (self.transform == other.transform)
-            and self.data.is_equal(other.data)
-        )
-
-
-def download_image(url: str, cache_dir: str):
-    parse = urlparse(url)
-    local_path = os.path.join(cache_dir, parse.netloc + parse.path)
-    if not os.path.exists(local_path):
-        os.makedirs(os.path.dirname(local_path), exist_ok=True)
-        try:
-            urllib.request.urlretrieve(url, local_path)
-        except (HTTPError, ConnectionResetError):
-            logger.warning(f"Could not download {url}. Skipping.")
-            return None
-
-    return folder.default_loader(local_path)
-
-
-class Downloader:
-    def __init__(
-        self,
-        cache_dir: str,
-        downloader: callable = None,
-    ):
-        self.cache_dir = cache_dir
-        if not os.path.exists(cache_dir):
-            os.makedirs(cache_dir)
-
-        if downloader is None:
-            self.downloader = download_image
-
-    def __call__(self, url: str):
-        return self.downloader(url, self.cache_dir)
