+++ meerkat/columns/lambda_column.py
@@ -1,11 +1,9 @@
 from __future__ import annotations
 
-import base64
 import logging
 import os
 import warnings
-from io import BytesIO
-from typing import Collection, Mapping, Sequence, Union
+from typing import Callable, Collection, Mapping, Sequence, Union
 
 import numpy as np
 import yaml
@@ -14,6 +12,7 @@ import meerkat as mk
 from meerkat.cells.abstract import AbstractCell
 from meerkat.columns.abstract import AbstractColumn
 from meerkat.datapanel import DataPanel
+from meerkat.display import lambda_cell_formatter
 from meerkat.errors import ConcatWarning
 from meerkat.tools.lazy_loader import LazyLoader
 
@@ -161,27 +160,9 @@ class LambdaColumn(AbstractColumn):
         else:
             return DataPanel.read(os.path.join(path, "data"))
 
+    @staticmethod
+    def _get_default_formatter() -> Callable:
+        return lambda_cell_formatter
+
     def _repr_cell(self, idx):
         return self.lz[idx]
-
-    def _get_formatter(self) -> callable:
-        if not mk.config.DisplayOptions.show_images:
-            return None
-
-        max_image_width = mk.config.DisplayOptions.max_image_width
-        max_image_height = mk.config.DisplayOptions.max_image_height
-
-        def _image_base64(im):
-            with BytesIO() as buffer:
-                im.save(buffer, "jpeg")
-                return base64.b64encode(buffer.getvalue()).decode()
-
-        def _image_formatter(cell):
-            im = cell.get()
-            if isinstance(im, Image.Image):
-                im.thumbnail((max_image_width, max_image_height))
-                return f'<img src="data:image/jpeg;base64,{_image_base64(im)}">'
-            else:
-                return repr(cell)
-
-        return _image_formatter
