+++ tests/meerkat/test_display.py
@@ -1,0 +1,3 @@
+# TODO
+
+# Test display oiptions
